transferring to implement CRC

+++ Core/Inc/app_comm.h
@@ -7,7 +7,7 @@
 
 // -- DEBUG -- //
 void sensorInit(UART_HandleTypeDef *huart);
-void createMetadataString(char *buf, size_t buf_size, uint8_t string_count, uint8_t pool_tail);
+void createMetadataString(char *buf, size_t buf_size, uint8_t string_count, uint16_t pool_tail);
 
 bool add(char *message, size_t len);
 void rem(void);
